adk-summarizer/requirements.txt
flask==3.0.2
gunicorn==21.2.0
google-generativeai